refactor(stoa-connect): split webMethods adapter into cohesive files
Split stoa-go/internal/connect/adapters/webmethods.go (1407 LOC) and
its test file (1489 LOC) into 13 source files + 12 test files, grouped
by responsibility: adapter, types, http, spec, lifecycle, discover,
sync, policy, credentials, oidc, alias, telemetry, config.

Behavior-preserving — public API surface unchanged (GatewayAdapter,
OIDCAdapter, AliasAdapter interfaces; NewWebMethodsAdapter constructor;
GetFailedRoutes, NormalizeEvent, HandleWebhookEvents exported symbols).
Callers in internal/connect/discovery.go and routes.go compile without
edits.

Extracted two helpers out of SyncRoutes: buildSyncPayload (spec-aware
payload builder) and parseCreateResponseID (nested/flat response parser).
Core SyncRoutes loop drops from ~165 LOC to ~95 LOC.

All 42 pre-existing tests preserved and green. Every webmethods_*.go
file is under 300 LOC (largest: policy_test.go at 291).
