+++ stoa-go/internal/connect/adapters/webmethods_adapter_test.go
@@ -1,0 +1,46 @@
+package adapters
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+)
+
+func TestWebMethodsDetect(t *testing.T) {
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/rest/apigateway/health" {
+			w.Header().Set("Content-Type", "application/json")
+			_ = json.NewEncoder(w).Encode(map[string]string{"status": "ok"})
+			return
+		}
+		w.WriteHeader(http.StatusNotFound)
+	}))
+	defer server.Close()
+
+	adapter := NewWebMethodsAdapter(AdapterConfig{Username: "Administrator", Password: "manage"})
+	ok, err := adapter.Detect(context.Background(), server.URL)
+	if err != nil {
+		t.Fatalf("detect error: %v", err)
+	}
+	if !ok {
+		t.Error("expected webMethods to be detected")
+	}
+}
+
+func TestWebMethodsDetectNotWebMethods(t *testing.T) {
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusNotFound)
+	}))
+	defer server.Close()
+
+	adapter := NewWebMethodsAdapter(AdapterConfig{})
+	ok, err := adapter.Detect(context.Background(), server.URL)
+	if err != nil {
+		t.Fatalf("detect error: %v", err)
+	}
+	if ok {
+		t.Error("expected webMethods NOT to be detected")
+	}
+}
